+++ host/include/backend/fpga/fpga.h
@@ -39,7 +39,7 @@ namespace AdapChol {
         RunPtr *runs;
         bool **P;
         int cus = 1;
-        int64_t waitTimeCount = 0, fillPTimeCount = 0, LeafCPUTimeCount = 0,
+        int waitTimeCount = 0, fillPTimeCount = 0, LeafCPUTimeCount = 0,
                 getFMemTimeCount = 0, syncTimeCount = 0, firstColProcTimeCount = 0,
                 preProcessAMatrixTimeCount = 0, returnFMemTimeCount = 0, kernelConstructRunTimeCount = 0,
                 rootNodeTimeCount = 0, argSetTimeCount = 0;
@@ -65,7 +65,7 @@ namespace AdapChol {
 
         bool *allocateP(size_t bytes) override;
 
-        int64_t getTimeCount() override;
+        int getTimeCount() override;
 
         void printStatistics() override;
